hw11 done

+++ src/src/main/java/hw11/MainApp.java
@@ -27,6 +27,7 @@ public class MainApp {
         String dobM = sc.nextLine();
         System.out.print("Введи iq матери: ");
         int iqM = sc.nextInt();
+        if(motherName == null || motherSurname == null || dobM == null) throw new IllegalArgumentException();
         Woman mother = new Woman(motherName, motherSurname, dobM, iqM);
         String buffer = sc.nextLine();
         System.out.print("Введи имя отца: ");
@@ -37,27 +38,34 @@ public class MainApp {
         String dobF = sc.nextLine();
         System.out.print("Введи iq отца: ");
         int iqF = sc.nextInt();
+        if(fatherName == null || fatherSurname == null || dobF == null) throw new IllegalArgumentException();
         Man father = new Man(fatherName, fatherSurname, dobF, iqF);
         FC.createNewFamily(mother, father);
     }
-    public static void makeABaby(){
+
+    public static void makeABaby() {
         Scanner sc = new Scanner(System.in);
         System.out.print("Введи порядковый номер семьи: ");
         int index = sc.nextInt();
+        if(index < 0) throw new IllegalArgumentException();
         Family family = FC.getFamilyByIndex(index - 1);
+        if(family.getChildrenQnty() > 2) throw new FamilyOverFlownException("Maybe it is enough?");
         String buffer = sc.nextLine();
         System.out.print("Какое имя дать мальчику? ");
         String mName = sc.nextLine();
         System.out.print("А девочке? ");
         String fName = sc.nextLine();
-//        sc.close();
-        FC.bornChild(family,mName,fName);
+        if(mName == null || fName == null) throw new IllegalArgumentException();
+        FC.bornChild(family, mName, fName);
     }
-    public static void adoptChild(){
+
+    public static void adoptChild() {
         Scanner sc = new Scanner(System.in);
         System.out.print("Введи номер семьи: ");
         int index = sc.nextInt();
+        if(index < 0) throw new IllegalArgumentException();
         Family f = FC.getFamilyByIndex(index - 1);
+        if(f.getChildrenQnty() > 2) throw new FamilyOverFlownException("Maybe it is enough?");
         String buffer = sc.nextLine();
         System.out.print("Введи имя ребенка: ");
         String name = sc.nextLine();
@@ -67,8 +75,8 @@ public class MainApp {
         String dob = sc.nextLine();
         System.out.print("Введи iq ребенка: ");
         int iq = sc.nextInt();
+        if(name == null || surname == null || dob == null) throw new IllegalArgumentException();
         Human kid = new Human(name, surname, dob, iq);
-//        sc.close();
         FC.adoptChild(f, kid);
     }
 
@@ -89,16 +97,19 @@ public class MainApp {
                 case 3:
                     System.out.print("Введи количество:");
                     int numberMoreThan = sc.nextInt();
+                    if(numberMoreThan < 2) throw new IllegalArgumentException();
                     FC.getFamiliesBiggerThan(numberMoreThan).forEach(System.out::println);
                     break;
                 case 4:
                     System.out.print("Введи количество:");
                     int numberLessThan = sc.nextInt();
+                    if(numberLessThan < 2) throw new IllegalArgumentException();
                     FC.getFamiliesLessThan(numberLessThan).forEach(System.out::println);
                     break;
                 case 5:
                     System.out.print("Введи количество:");
                     int exactNumber = sc.nextInt();
+                    if(exactNumber < 2) throw new IllegalArgumentException();
                     int i = FC.countFamiliesWithMemberNumber(exactNumber);
                     System.out.printf("Всего %d семей \n", i);
                     break;
@@ -108,6 +119,7 @@ public class MainApp {
                 case 7:
                     System.out.print("Введи порядковый номер:");
                     int familyNumber = sc.nextInt();
+                    if(familyNumber < 0) throw new IllegalArgumentException();
                     FC.deleteFamilyByIndex(familyNumber - 1);
                     break;
                 case 8:
@@ -116,14 +128,15 @@ public class MainApp {
                     System.out.println("3. Вернуться назад");
                     System.out.print("Что делаем? ");
                     int option = sc.nextInt();
-                    switch (option){
+                    switch (option) {
                         case 1:
                             makeABaby();
                             break;
                         case 2:
                             adoptChild();
                             break;
-                        case 3: break;
+                        case 3:
+                            break;
                         default:
                             throw new IllegalStateException("Unexpected value: " + option);
                     }
@@ -131,6 +144,7 @@ public class MainApp {
                 case 9:
                     System.out.print("Введи возраст: ");
                     int age = sc.nextInt();
+                    if(age < 1) throw new IllegalArgumentException();
                     FC.deleteAllChildrenOlderThen(age);
                     break;
                 case 0:
